@@ -4,10 +4,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build:prod": "webpack -p",
-    "build": "webpack",
+    "build:prod": "webpack -p --mode=production",
+    "build": "webpack -d --mode=development",
     "watch": "webpack -w",
-    "dev": "webpack-dev-server --hot --inline",
+    "dev": "webpack-dev-server --hot --inline --mode=development",
     "start": "npm run dev",
     "test": "jest",
     "test:watch": "jest --watch"
@@ -23,31 +23,36 @@
     "url": "https://github.com/dotKom/glowing-fortnight/issues"
   },
   "homepage": "https://github.com/dotKom/glowing-fortnight#readme",
+  "jest": {
+    "setupTestFrameworkScriptFile": "<rootDir>/components/testSetup.js"
+  },
   "devDependencies": {
-    "babel-core": "^6.13.2",
-    "babel-jest": "^20.0.3",
-    "babel-loader": "^6.2.4",
+    "babel-core": "^6.26.3",
+    "babel-jest": "^23.2.0",
+    "babel-loader": "^7.1.4",
     "babel-plugin-transform-react-require": "^1.0.1",
-    "babel-preset-es2015": "^6.13.2",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.11.1",
-    "css-loader": "^0.23.1",
-    "enzyme": "^2.9.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "jest": "^20.0.4",
-    "node-sass": "^3.8.0",
-    "react-test-renderer": "^15.6.1",
-    "sass-loader": "^4.0.0",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "css-loader": "^0.28.11",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "jest": "^23.2.0",
+    "mini-css-extract-plugin": "^0.4.1",
+    "node-sass": "^4.9.0",
+    "react-test-renderer": "^16.4.1",
+    "sass-loader": "^7.0.3",
+    "style-loader": "^0.21.0",
+    "webpack": "^4.14.0",
+    "webpack-cli": "^3.0.8",
+    "webpack-dev-server": "^3.1.4"
   },
   "dependencies": {
-    "babel-polyfill": "^6.13.0",
-    "es6-promise": "^3.2.1",
+    "babel-polyfill": "^6.26.0",
+    "es6-promise": "^4.2.4",
     "isomorphic-fetch": "^2.2.1",
-    "moment": "^2.18.1",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0",
+    "moment": "^2.22.2",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
     "remarkable": "^1.6.2"
   }
 }
